+++ src/components/ADempiere/Form/VPOS/Order/orderLineMixin.js
@@ -83,7 +83,6 @@ export default {
     },
     listOrderLines({ uuid: orderUuid }) {
       if (!this.isEmptyValue(orderUuid)) {
-        this.$store.dispatch('listOrderLinesFromServer', orderUuid)
         this.orderLines = this.listOrderLine
         this.handleCurrentLineChange(this.currentOrderLine)
       }
@@ -152,8 +151,7 @@ export default {
       if (columnName === 'LineDescription') {
         return row.lineDescription
       }
-
-      const currency = this.currencyPoint.iSOCode
+      const currency = this.currencyPointOfSales.iSOCode
       if (columnName === 'CurrentPrice') {
         return this.formatPrice(row.priceActual, currency)
       } else if (columnName === 'QtyOrdered') {
