+++ src/store/modules/ADempiere/pointOfSales/keyLayout/index.js
@@ -1,0 +1,18 @@
+
+import state from './state.js'
+import mutations from './mutations.js'
+import actions from './actions.js'
+import getters from './getters.js'
+
+/**
+ * keyLayout Vuex Module
+ * @author Elsio Sanchez <elsiosanches@gmail.com>
+ */
+const keyLayout = {
+  state,
+  mutations,
+  actions,
+  getters
+}
+
+export default keyLayout
