+++ src/store/modules/ADempiere/pointOfSales/keyLayout/state.js
@@ -1,0 +1,7 @@
+/**
+ * keyLayout State
+ * @author Elsio Sanchez <elsiosanches@gmail.com>
+ */
+export default {
+  keyLayout: {}
+}
