+++ src/store/modules/ADempiere/pointOfSales/point/index.js
@@ -1,0 +1,18 @@
+
+import state from './state.js'
+import mutations from './mutations.js'
+import actions from './actions.js'
+import getters from '../getters/index.js'
+
+/**
+ * Pos Vuex Module
+ * @author Elsio Sanchez <elsiosanches@gmail.com>
+ */
+const pointOfSales = {
+  state,
+  mutations,
+  actions,
+  getters
+}
+
+export default pointOfSales
