+++ src/store/modules/ADempiere/pointOfSales/point/state.js
@@ -1,0 +1,17 @@
+/**
+ * Pos State
+ * @author Elsio Sanchez <elsiosanches@gmail.com>
+ */
+const withoutResponse = {
+  isLoaded: false,
+  isReload: true,
+  recordCount: 0,
+  nextPageToken: undefined
+}
+export default {
+  pointOfSales: {
+    ...withoutResponse
+  },
+  listPointOfSales: {},
+  currentPointOfSales: {}
+}
