+++ src/store/modules/ADempiere/utils/PointOfSales/getters.js
@@ -1,0 +1,16 @@
+/**
+ * PointOfSales Getters
+ * @author Elsio Sanchez <elsiosanches@gmail.com>
+ */
+
+export default {
+  getIsShowPOSOptions: (state) => {
+    return state.showPOSOptions
+  },
+  getShowPOSKeyLayout: (state) => {
+    return state.showPOSKeyLayout
+  },
+  getShowCollectionPos: (state) => {
+    return state.showPOSCollection
+  }
+}
