+++ src/store/modules/ADempiere/utils/PointOfSales/index.js
@@ -1,0 +1,16 @@
+
+import state from './state.js'
+import mutations from './mutations.js'
+import getters from './getters.js'
+
+/**
+ * Pos Vuex Module
+ * @author Elsio Sanchez <elsiosanches@gmail.com>
+ */
+const pointOfSalesUtils = {
+  state,
+  mutations,
+  getters
+}
+
+export default pointOfSalesUtils
