+++ src/store/modules/ADempiere/utils/PointOfSales/mutations.js
@@ -1,0 +1,15 @@
+/**
+ * Pos Mutations
+ * @author Elsio Sanchez <elsiosanches@gmail.com>
+ */
+export default {
+  setShowPOSOptions(state, isShowedOptions) {
+    state.showPOSOptions = isShowedOptions
+  },
+  setShowPOSKeyLayout(state, isShowedKeyLayout) {
+    state.showPOSKeyLayout = isShowedKeyLayout
+  },
+  setShowPOSCollection(state, isShowedCollection) {
+    state.showPOSCollection = isShowedCollection
+  }
+}
