+++ src/store/modules/ADempiere/utils/PointOfSales/state.js
@@ -1,0 +1,9 @@
+/**
+ * Pos State
+ * @author Elsio Sanchez <elsiosanches@gmail.com>
+ */
+export default {
+  showPOSOptions: false,
+  showPOSKeyLayout: false,
+  showPOSCollection: false
+}
